Move pump traits to it's own namespace
