+++ src/Library/Benzina/Pump/UsersPump.php
@@ -4,6 +4,8 @@ namespace App\Library\Benzina\Pump;
 
 use App\Entity\Accounting;
 use App\Entity\User;
+use App\Library\Benzina\Pump\Trait\ArrayPumpTrait;
+use App\Library\Benzina\Pump\Trait\ProgressivePumpTrait;
 use Doctrine\ORM\EntityManagerInterface;
 
 class UsersPump implements PumpInterface
